+++ lib/mobius/auto_save.ex
@@ -7,18 +7,18 @@ defmodule Mobius.AutoSave do
 
   @spec start_link([Mobius.arg()]) :: GenServer.on_start()
   def start_link(args) do
-    GenServer.start_link(__MODULE__, args, name: name(args[:name]))
+    GenServer.start_link(__MODULE__, args, name: name(args[:mobius_instance]))
   end
 
-  defp name(mobius_name) do
-    Module.concat(__MODULE__, mobius_name)
+  defp name(instance) do
+    Module.concat(__MODULE__, instance)
   end
 
   @impl GenServer
   def init(args) do
     state =
       args
-      |> Keyword.take([:autosave_interval, :name, :persistence_dir])
+      |> Keyword.take([:autosave_interval, :mobius_instance, :persistence_dir])
       |> Enum.into(%{})
 
     _ = :timer.send_interval(state.autosave_interval * 1_000, self(), :auto_save)
@@ -28,7 +28,7 @@ defmodule Mobius.AutoSave do
 
   @impl GenServer
   def handle_info(:auto_save, state) do
-    _ = Mobius.save(state.name)
+    _ = Mobius.save(state.mobius_instance)
     {:noreply, state}
   end
 end